made untiling layer, but unsure how to best apply it.
Merging properties of dat.Data object with properties of models, which shouldn't
have to be aware of actual image shape, ideally, just the tile shape. But
sometimes they need to be, in order to untile. Should models also do the tiling?
Seemingly not, since we may want to shuffle tiles eventually, and you can't
parallelize within the model.