wrote functions for skip, take, and split, but this is probably dumb
Need to store train and test sets separately I think. Can do this in
convert. But don't call them different things, just "labeled_0-1699.tfrecord"
for example, and then we can pass in the name of "train-set" and "test-set"?
Maybe? Maybe not. This has the disadvantage of having a bunch of disjoint sets,
poorly named, in the data-dir.

Could look in "labeled" directory for labeled sets and "test" directory for
test sets and just glob all the tfrecords there. That would allow an annotator
later on to grab those every epoch. For now, it lets us separate training and
test sets definitively, with none of this business of skipping and taking, which
isn't extensible down the road when we have lots of unlabeled data, want to
store annotated labeled_sets separately and load them together.

Unclear. need to figure this out...