wrote some code for multi-scale tracking
need to then use the peak footprints predicted at each tile to search for peaks
in the full, untiled image, in order to deal with edge cases. If maximum was
found at an edge, then just have to make sure the footprint around it includes
the real peak.

This should be guaranteed, as long as we've tuned threshold_abs corectly.